refactor(app): extract usePubtatorAdminPanel from ManagePubtator (#399)
Reduce ManagePubtator.vue from 638 to 548 lines by moving the view-level
form state, feedback banner, clear-all modal flag, cacheStateLabel,
formatDate, and the four action handlers into a usePubtatorAdminPanel
composable that wraps the existing usePubtatorAdmin cache/job composable.

The view is now a thin template + single composable call. Behavior,
network calls, and the rendered structure are unchanged.

Advances #399. Part of #346.
